pyproject.toml
[tool.poetry]
name = "jwt-proxy"
version = "0.1.0"
description = ""
authors = ["Sebastian Saldarriaga <sebastian.saldarriaga.ramirez@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.65.2"
uvicorn = "^0.14.0"
pydantic = "^1.8.2"
pyjwt = "^2.1.0"
httpx = "^0.18.2"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.9"
requests = "^2.25.1"
pytest-testdox = "^2.0.1"
pytest = "^6.2.4"
pytest_httpx = "^0.12.0"
mock = "^4.0.3"
pytest-asyncio = "^0.15.1"
pytest-cov = "^2.12.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
